# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
...
  /api/v1/products/{productId}/image:
    get:
      tags:
      - Images
      summary: Get product image metadata
      description: Retrieves image metadata for a specific product
      operationId: getProductImage
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Product or image not found
        "200":
          description: Image metadata retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
    post:
      tags:
      - Images
      summary: Replace product image
      description: Uploads a new image for a specific product and replaces the existing
        one
      operationId: replaceProductImage
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "404":
          description: Product not found
        "201":
          description: Image replaced successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "400":
          description: Invalid input or image processing failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized user
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/inquiries:
    post:
      tags:
      - Inquiries
      summary: Submit inquiry
      description: Submits a new inquiry from a buyer to a seller
      operationId: sendInquiry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Inquiry created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResultDTO"
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResultDTO"
        "401":
          description: Unauthorized user
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResultDTO"
        "409":
          description: "Inquiry could not be created (e.g., cooldown active)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResultDTO"
...
  /api/v1/images/{imageId}/file:
    get:
      tags:
      - Images
      summary: Get image file
      description: Retrieves the actual image file bytes as a JPEG
      operationId: getImageFile
      parameters:
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Image not found
        "200":
          description: Image file retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "500":
          description: Database error while retrieving the image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    UserDTO:
      type: object
      description: Represents a user account with public profile and authentication
        role information
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
          example: 5
        name:
          type: string
          description: User's display name
          example: Laura Martinez
        email:
          type: string
          description: User's email address
          example: laura.martinez@example.com
        rating:
          type: number
          format: double
          description: Average rating received by the user as a seller
          example: 4.8
        numRatings:
          type: integer
          format: int32
          description: Total number of ratings received
          example: 34
        description:
          type: string
          description: User's profile description or biography
          example: Seller of vintage photography items.
        roles:
          type: array
          description: "List of assigned security roles (e.g., ROLE_USER, ROLE_ADMIN)"
          items:
            type: string
        banned:
          type: boolean
          description: Boolean indicating if the user account is banned
          example: false
...
    ImageDTO:
      type: object
      description: Represents a product image reference in the system
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the image
          example: 101
    ProductDTO:
      type: object
      description: "Represents a product with its core details, seller information,\
        \ and user interactions"
      properties:
        id:
          type: integer
          format: int64
          description: Unique product identifier
          example: 1
        name:
          type: string
          description: Product name or title
          example: Mountain Bike
        category:
          type: string
          description: Product category classification
          example: Sports
        price:
          type: number
          format: double
          description: Current selling price of the product
          example: 120.5
        location:
          type: string
          description: Geographic location where the product is listed
          example: Madrid
        description:
          type: string
          description: Detailed product description
          example: "Mountain bike almost new, very little use."
        status:
          type: string
          description: "Product status (e.g., available, sold, inactive)"
          example: available
        image:
          $ref: "#/components/schemas/ImageDTO"
          description: Product image reference
        seller:
          $ref: "#/components/schemas/UserDTO"
          description: Information about the product seller
        userInteractions:
          type: array
          description: "List of user interactions (views, purchases, etc.) on this\
            \ product"
          items:
            $ref: "#/components/schemas/UserInteractionDTO"
...
    UserInteractionDTO:
      type: object
      description: "Represents a user interaction event with a product (view, purchase,\
        \ etc.)"
      properties:
        id:
          type: integer
          format: int64
          description: Unique interaction identifier
          example: 204
        userId:
          type: integer
          format: int64
          description: ID of the user who performed the interaction
          example: 12
        type:
          type: string
          description: "Type of interaction (e.g., VIEW, PURCHASE)"
          enum:
          - VIEW
          - LIKE
          - BUY
          example: VIEW
...
    InquiryRequestDTO:
      type: object
      description: Request body for creating a new customer inquiry
      properties:
        productId:
          type: integer
          format: int64
          description: ID of the product being inquired about
          example: 12
        phone:
          type: string
          description: Customer's phone number
          example: "+34611223344"
        type:
          type: string
          description: Type or category of the inquiry
          example: Product status
        message:
          type: string
          description: The inquiry message content
          example: Does the screen have any scratches?
    InquiryDTO:
      type: object
      description: Represents a customer inquiry about a product with contact and
        product information
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the inquiry
          example: 50
        productName:
          type: string
          description: Name of the product being inquired about
          example: Canon Reflex Camera
        sellerEmail:
          type: string
          description: Email address of the product seller
          example: seller@example.com
        buyerName:
          type: string
          description: Name of the customer making the inquiry
          example: Ana Lopez
        buyerEmail:
          type: string
          description: Email address of the customer
          example: ana.lopez@example.com
        buyerPhone:
          type: string
          description: Phone number of the customer
          example: "+34600123456"
        inquiryType:
          type: string
          description: Type or category of the inquiry
          example: Question about delivery
        message:
          type: string
          description: The inquiry message content
          example: Do you ship to the Canary Islands?
          maxLength: 1000
          minLength: 5
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the inquiry was created
          example: 2024-03-24T10:15:30
        status:
          type: string
          description: "Current status of the inquiry (e.g., pending, answered)"
          example: pending
        product:
          $ref: "#/components/schemas/ProductDTO"
          description: The product being inquired about
        buyer:
          $ref: "#/components/schemas/UserDTO"
          description: The customer making the inquiry
      required:
      - message
      - status
    NotificationResultDTO:
      type: object
      description: Result of attempting to send a notification about an inquiry
      properties:
        inquiry:
          $ref: "#/components/schemas/InquiryDTO"
          description: The inquiry that notification was attempted for
        sent:
          type: boolean
          description: Boolean indicating if the notification was successfully sent
          example: true
        cooldownMinutes:
          type: integer
          format: int64
          description: Minutes remaining before another notification can be sent (if
            cooldown is active)
          example: 0
        errorCode:
          type: string
          description: "Error code if notification failed, null if successful"
          example: null
...